requirements.txt
pyTelegramBotApi==3.6.7
Pyyaml==6.0
